fix: 클래스명 통일 & ClubAdmin -> ClubManager으로 변경
