+++ src/main/java/in/koreatech/koin/admin/club/dto/request/AdminChangeClubActiveRequest.java
@@ -9,7 +9,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
-public record ChangeAdminClubActiveRequest(
+public record AdminChangeClubActiveRequest(
     @NotNull
     @Schema(description = "동아리 활성화 여부", example = "false", requiredMode = REQUIRED)
     Boolean isActive
