+++ src/main/java/in/koreatech/koin/admin/club/dto/request/AdminCreateClubRequest.java
@@ -9,13 +9,14 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 import in.koreatech.koin.domain.club.model.Club;
 import in.koreatech.koin.domain.club.model.ClubManager;
+import in.koreatech.koin.domain.club.model.ClubCategory;
 import in.koreatech.koin.domain.user.model.User;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
-public record ModifyAdminClubRequest(
+public record AdminCreateClubRequest(
     @Schema(description = "동아리 이름", example = "BCSD Lab", requiredMode = REQUIRED)
     @NotEmpty(message = "동아리 이름은 필수 입력 사항입니다.")
     String name,
@@ -24,13 +25,13 @@ public record ModifyAdminClubRequest(
     @NotNull(message = "동아리 사진은 필수 입력 사항입니다.")
     String imageUrl,
 
+    @Schema(description = "동아리 관리자 ID 리스트", requiredMode = REQUIRED)
+    @NotEmpty(message = "동아리 관리자는 필수 입력 사항입니다.")
+    List<InnerClubManagerRequest> clubManagers,
+
     @Schema(description = "동아리 분과 카테고리 ID", example = "1", requiredMode = REQUIRED)
     @NotNull(message = "동아리 분과 카테고리는 필수 입력 사항입니다.")
     Integer clubCategoryId,
-
-    @Schema(description = "동아리 관리자 ID 리스트", requiredMode = REQUIRED)
-    @NotEmpty(message = "동아리 관리자는 필수 입력 사항입니다.")
-    List<InnerClubManagerUpdateRequest> clubManagers,
 
     @Schema(description = "동아리 위치", example = "학생회관", requiredMode = REQUIRED)
     @NotNull(message = "동아리 위치는 필수 입력 사항입니다.")
@@ -39,10 +40,6 @@ public record ModifyAdminClubRequest(
     @Schema(description = "동아리 소개", example = "즐겁게 일하고 열심히 노는 IT 특성화 동아리", requiredMode = REQUIRED)
     @NotNull(message = "동아리 소개는 필수 입력 사항입니다.")
     String description,
-
-    @Schema(description = "동아리 활성화 여부", example = "false", requiredMode = REQUIRED)
-    @NotNull(message = "동아리 활성화 여부는 필수 입력 사항입니다.")
-    Boolean active,
 
     @Schema(description = "인스타그램 링크", example = "https://www.instagram.com/bcsdlab/", requiredMode = REQUIRED)
     String instagram,
@@ -57,7 +54,7 @@ public record ModifyAdminClubRequest(
     String phoneNumber
 ) {
     @JsonNaming(value = SnakeCaseStrategy.class)
-    public record InnerClubManagerUpdateRequest(
+    public record InnerClubManagerRequest(
         @Schema(description = "동아리 관리자 id", example = "bcsdlab", requiredMode = REQUIRED)
         String userId
     ) {
@@ -67,5 +64,20 @@ public record ModifyAdminClubRequest(
                 .user(user)
                 .build();
         }
+    }
+
+    public Club toEntity(ClubCategory clubCategory) {
+        return Club.builder()
+            .name(name)
+            .lastWeekHits(0)
+            .active(false)
+            .likes(0)
+            .hits(0)
+            .introduction("")
+            .imageUrl(imageUrl)
+            .clubCategory(clubCategory)
+            .description(description)
+            .location(location)
+            .build();
     }
 }
