+++ src/main/java/in/koreatech/koin/admin/club/dto/request/AdminDecideClubManagerRequest.java
@@ -9,7 +9,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
-public record DecideClubManagerAdminRequest(
+public record AdminDecideClubManagerRequest(
     @NotNull
     @Schema(description = "동아리 이름", example = "BCSD", requiredMode = REQUIRED)
     String clubName,
