+++ src/main/java/in/koreatech/koin/admin/club/service/AdminClubService.java
@@ -17,11 +17,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import in.koreatech.koin._common.model.Criteria;
-import in.koreatech.koin.admin.club.dto.request.ClubManagerAdminCondition;
-import in.koreatech.koin.admin.club.dto.request.ChangeAdminClubActiveRequest;
-import in.koreatech.koin.admin.club.dto.request.CreateAdminClubRequest;
-import in.koreatech.koin.admin.club.dto.request.DecideClubManagerAdminRequest;
-import in.koreatech.koin.admin.club.dto.request.ModifyAdminClubRequest;
+import in.koreatech.koin.admin.club.dto.request.AdminClubManagerCondition;
+import in.koreatech.koin.admin.club.dto.request.AdminChangeClubActiveRequest;
+import in.koreatech.koin.admin.club.dto.request.AdminCreateClubRequest;
+import in.koreatech.koin.admin.club.dto.request.AdminDecideClubManagerRequest;
+import in.koreatech.koin.admin.club.dto.request.AdminModifyClubRequest;
 import in.koreatech.koin.admin.club.dto.response.AdminClubManagersResponse;
 import in.koreatech.koin.admin.club.dto.response.AdminClubResponse;
 import in.koreatech.koin.admin.club.dto.response.AdminClubsResponse;
@@ -100,7 +100,7 @@ public class AdminClubService {
     }
 
     @Transactional
-    public void createClub(CreateAdminClubRequest request) {
+    public void createClub(AdminCreateClubRequest request) {
         ClubCategory clubCategory = adminClubCategoryRepository.getById(request.clubCategoryId());
         Club club = adminClubRepository.save(request.toEntity(clubCategory));
 
@@ -125,7 +125,7 @@ public class AdminClubService {
     }
 
     @Transactional
-    public void modifyClub(Integer clubId, ModifyAdminClubRequest request) {
+    public void modifyClub(Integer clubId, AdminModifyClubRequest request) {
         ClubCategory clubCategory = adminClubCategoryRepository.getById(request.clubCategoryId());
         Club club = adminClubRepository.getById(clubId);
 
@@ -162,13 +162,13 @@ public class AdminClubService {
     }
 
     @Transactional
-    public void changeActive(Integer clubId, ChangeAdminClubActiveRequest request) {
+    public void changeActive(Integer clubId, AdminChangeClubActiveRequest request) {
         Club club = clubRepository.getById(clubId);
         club.updateActive(request.isActive());
     }
 
     @Transactional
-    public void decideClubAdmin(String clubName, DecideClubManagerAdminRequest request) {
+    public void decideClubAdmin(String clubName, AdminDecideClubManagerRequest request) {
         ClubCreateRedis createRequest = clubCreateRedisRepository.findById(clubName)
             .orElseThrow(() -> ClubNotFoundException.withDetail("신청한 동아리를 찾을 수 없습니다."));
 
@@ -179,7 +179,7 @@ public class AdminClubService {
         clubCreateRedisRepository.deleteById(clubName);
     }
 
-    public AdminClubManagersResponse getClubAdmins(ClubManagerAdminCondition condition) {
+    public AdminClubManagersResponse getClubAdmins(AdminClubManagerCondition condition) {
         int totalCount = clubManagerRepository.countAll();
         Criteria criteria = Criteria.of(condition.page(), condition.limit(), totalCount);
         Sort.Direction direction = condition.getDirection();
@@ -189,7 +189,7 @@ public class AdminClubService {
         return AdminClubManagersResponse.of(result, criteria);
     }
 
-    public AdminClubManagersResponse getUnacceptedClubManagers(ClubManagerAdminCondition condition) {
+    public AdminClubManagersResponse getUnacceptedClubManagers(AdminClubManagerCondition condition) {
         List<ClubCreateRedis> unAcceptedClubList = (List<ClubCreateRedis>)clubCreateRedisRepository.findAll();
         int totalCount = unAcceptedClubList.size();
         Criteria criteria = Criteria.of(condition.page(), condition.limit(), totalCount);
